refactor categories page
functionality check status button refactor
