+++ src/components/pages/categories/Categories.js
@@ -1,12 +1,12 @@
 import { useDispatch, useSelector } from 'react-redux/es/exports';
-import categoriesOfBooks from '../../../redux/categories/categories';
+import { checking } from '../../../redux/categories/categories';
 
 export default function Categories() {
   const dispatch = useDispatch();
   const categories = useSelector((state) => state.categories.categories);
 
   const checkStatusHandler = () => {
-    dispatch(categoriesOfBooks('Under construction'));
+    dispatch(checking('Under construction'));
   };
 
   return (
