--- src/redux/categoriesReducer.js
@@ -1,16 +1,0 @@
-import { createReducer } from '@reduxjs/toolkit';
-import categoriesOfBooks from './categories/categories';
-
-const initialState = {
-  categories: '',
-};
-
-const categoriesReducer = createReducer(initialState, (builder) => {
-  builder
-    .addCase(categoriesOfBooks, (state, action) => ({
-      ...state, categories: action.payload,
-    }))
-    .addDefaultCase((state) => state);
-});
-
-export default categoriesReducer;
