+++ src/redux/store/configureStore.js
@@ -1,6 +1,6 @@
 import { configureStore } from '@reduxjs/toolkit';
 import bookReducer from '../books/books';
-import categoriesReducer from '../categoriesReducer';
+import categoriesReducer from '../categories/categories';
 
 const rootReducer = {
   books: bookReducer.reducer,
